+++ DSCResources/MSFT_xSQLServerConfiguration/MSFT_xSQLServerConfiguration.psm1
@@ -35,8 +35,9 @@ function Get-TargetResource
         [String]
         $SQLServer,
 
+        [Parameter(Mandatory = $true)]
         [String]
-        $SQLInstanceName = 'MSSQLSERVER',
+        $SQLInstanceName,
 
         [Parameter(Mandatory = $true)]
         [String]
@@ -103,8 +104,9 @@ function Set-TargetResource
         [String]
         $SQLServer,
 
+        [Parameter(Mandatory = $true)]
         [String]
-        $SQLInstanceName = 'MSSQLSERVER',
+        $SQLInstanceName,
 
         [Parameter(Mandatory = $true)]
         [String]
@@ -182,8 +184,9 @@ function Test-TargetResource
         [String]
         $SQLServer,
 
+        [Parameter(Mandatory = $true)]
         [String]
-        $SQLInstanceName = 'MSSQLSERVER',
+        $SQLInstanceName,
 
         [Parameter(Mandatory = $true)]
         [String]