+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Changes to xSQLServerConfiguration
+  - BREAKING CHANGE: The parameter SQLInstanceName is now mandatory.
+  - Resource can now be used to define the configuration of two or more different DB instances on the same server.
+
 - Changes to xSQLServerRole
   - xSQLServerRole now correctly reports that the desired state is present when the login is already a member of the server roles